requirements.txt
absl-py==2.1.0
accelerate==0.23.0
addict==2.4.0
affine==2.4.0
aiofiles==24.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
anykeystore==0.2
apex==0.9.10.dev0
appdirs==1.4.4
attrs==21.4.0
beautifulsoup4==4.13.3
black==21.4b2
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.3.2
cityscapesScripts==2.2.2
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.1
coloredlogs==15.0.1
contourpy==1.3.1
cryptacular==1.6.2
cycler==0.12.1
Cython==3.0.2
deepspeed==0.18.0
defusedxml==0.7.1
detectron2==0.6
diffdist==0.1
docker-pycreds==0.4.0
easydict==1.13
einops==0.7.0
exceptiongroup==1.3.0
fastapi==0.118.3
ffmpy==0.6.2
filelock==3.13.1
fonttools==4.56.0
fsspec==2025.3.0
ftfy==6.1.1
future==1.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2==2.2.1
gradio==5.49.1
gradio_client==1.13.3
greenlet==3.1.1
groovy==0.1.2
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
hf-xet==1.1.10
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
hupper==1.12.1
hydra-core==1.3.2
idna==3.7
imageio==2.37.0
infinibatch==0.1.1
iopath==0.1.9
Jinja2==3.1.6
joblib==1.5.2
json-tricks==3.17.3
jsonlines==4.0.0
kiwisolver==1.4.8
kornia==0.6.8
lazy_loader==0.4
Levenshtein==0.27.1
llvmlite==0.45.1
Markdown==3.7
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
more-itertools==10.8.0
mpi4py==4.0.3
mpmath==1.3.0
msgpack==1.1.2
MultiScaleDeformableAttention==1.0
mup==1.0.0
mypy_extensions==1.1.0
networkx==3.4.2
ninja==1.13.0
nltk==3.8.1
numba==0.62.1
numpy==1.23.1
nvidia-ml-py==13.580.82
oauthlib==3.2.2
omegaconf==2.3.0
openai-whisper==20250625
opencv-python==4.8.1.78
orjson==3.11.3
packaging==24.2
pandas==2.0.3
PasteDeploy==3.1.0
pathspec==0.12.1
pathtools==0.1.2
pbkdf2==1.3
Pillow==9.4.0
pip==25.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.3.6
portalocker==3.2.0
protobuf==4.25.8
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==13.0.0
pycocotools==2.0.7
pydantic==2.10.6
pydantic_core==2.27.2
pydot==4.0.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.1
pyquaternion==0.9.9
pyramid==2.0.2
pyramid-mailer==0.15.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python3-openid==3.2.0
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.1
RapidFuzz==3.12.2
rasterio==1.3.0
regex==2023.10.3
repoze.sendmail==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.2.0
ruff==0.14.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.23.1
setproctitle==1.3.5
setuptools==75.8.0
Shapely==1.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snuggs==1.4.7
soupsieve==2.6
SQLAlchemy==2.0.39
starlette==0.48.0
supervision==0.25.1
sympy==1.13.3
tabulate==0.9.0
task-adapter==0.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
timm==1.0.20
tokenizers==0.21.1
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transaction==5.0
transformers==4.49.0
translationstring==1.4
triton==2.2.0
typer==0.19.2
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.37.0
velruse==1.1.1
venusian==3.1.1
vision-datasets==0.2.2
wandb==0.15.12
wcwidth==0.2.14
WebOb==1.8.9
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
wtforms-recaptcha==0.3.2
yacs==0.1.8
yapf==0.43.0